requirements.txt
beautifulsoup4==4.9.1
Django==3.0.8
django-bootstrap4==2.2.0
django-summernote==0.8.11.6
django-taggit==1.3.0
Pillow==7.2.0
pytz==2020.1
soupsieve==2.0.1

